requirements.txt
python-dotenv==0.19.2
psycopg2==2.9.3
django==3.2.12
djangorestframework=3.13.1
kafka-python=2.0.2